bhf: arm:zynqmp: Enable MMU and cache very early on ZynqMP
Enable MMU and caches very early in the boot process on ZynqMP in U-Boot
proper. This allows board_init_f to run with icache and dcache enabled,
which saves some 1800 milliseconds of boot time on ZynqMP based device.

Only for the first/low memory region (start adress:0x00, size: 2GB) of
the ZynqMP the MMU and the cache are enabled.

This is similar to:
ac9153c74f3 ("ARM: imx: Enable MMU and dcache very early on i.MX8M")