board: beckhoff: cx8200: debian: add reset button support
The Beckhoff CX8200 is equipped with a reset button under the front flap.
If this button is pressed, a file with the name "reset-button-pressed"
and the content "true" is located in the "chosen" folder of the device tree
later in the Linux operating system. This file can be evaluated by software
for different purposes.