board: beckhoff: add CX8200
Product description (From [1]):
The CX8200 Embedded PC features an Arm® Cortex®-A53 dual-core processor running at 1.2 GHz.
The basic configuration of the CX8200 includes a 1 Gbit Ethernet interface, a USB 3.0 interface,
a 1 GB LPDDR4 RAM, and a slot for a microSD card.

From u-boots view this is a pretty standard zynqmp based device. Compared
to the zynqmp defconfig some features are disabled and there's automatic
loading of fpga files based on eeprom data. We use u-boot spl boot from
sdcard, so there is no fsbl on the device. We also need to adjust the
fdt_addr_r variable so we fit inside the devices 1GB of memory

[1] https://www.beckhoff.com/en-en/products/ipc/embedded-pcs/cx8200-arm-r-cortex-r-a53/cx8200.html