board: beckhoff: add CX9240
Hardware features from the product description [1]:

- high-performance 1.2 GHz Arm® Cortex®-A53 CPUs with four cores
- 2 GB LPDDR4 RAM
- two independent 1 Gbit Ethernet interfaces
- DisplayPort instead of DVI
- built-in capacitive 1-second UPS
- electrically isolated 24 V DC power supply
- four USB 3.0 interfaces
- a DisplayPort,
- a multi-option interface that can be flexibly assigned ex factory – for example, with a serial port or a fieldbus connection.

[1]: https://www.beckhoff.com/en-en/products/ipc/embedded-pcs/cx9240-arm-r-cortex-r-a53/cx9240.html

From u-boots view this device is very similar to the previously
added CX8200. The main difference is the display port connection. We
don't  enable it in u-boot as using the display requires loading an fpga
file first and we currently do that fairly late in boot.