bhf: add ci and debian package build
This is a squashed version of all the packaging / ci related changes
from our bhf/v2024.10 branch. For git history look there.

We create a debian package for our CX8200 / 9240 devices that extracts
the u-boot files and fpga images to our efi partition. Since the efi is
fat, apt won't install there directly and we copy the files.

We also add a service to initialize the rtc configuration to use the
battery.